mark objects with the cacheAt attribute when they arrived in the cached && fix the controller

+++ burst/coldStorageController.py
@@ -14,52 +14,72 @@
 
 import argparse
 import pyh3lib
+import time
 
-def MoveToCold(h3_cache, h3_cold):
+DEFAULT_EXPIRES_AT_TIME = 1800 
+"""Maximum default time that the objects stay in the cache (30 minutes)"""
+
+def MoveToCold(h3_cache, h3_cold, expires_time):
     """
     Moves the objects from the the cache storage to 
     the cold storage
 
     :param h3_cache: h3 cache storage
     :param h3_cold: h3 cold storage
+    :param now : the "now" time
     :type h3_cache: object
     :type h3_cold: object
+    :type now: float
     :returns: nothing
     """
+
+    now = time.get_time(time.CLOCK_REALTIME)
 
     # List all the buckets in the cache storage
     for h3_cache_bucket in h3_cache.list_buckets():
         # List all the objects in the cache bucket
         for h3_cache_object in h3_cache.list_objects(h3_cache_bucket):
-            done   = False
-            offset = 0
-            
-            while not done:
-                data = h3_cache.read_object(h3_cache_bucket, h3_cache_object, offset)
-                h3_cold.write_object(h3_cache_bucket, h3_cache_object, data, offset)
-                
-                done    = data.done
-                offset += len(data)
+            # get when it expires from the cache    
+            cached_at = struct.unpack('d', h3_cold.read_object_metadata(h3_cache_bucket, h3_cache_object, "CachedAt"))
 
-            h3_cache.delete_object(h3_cache_bucket, h3_cache_object)
+            # the object must be moved back to the cold storage
+            if cached_at and (cached_at[0] + expires_time) >= now:
+                done   = False
+                offset = 0
+                
+                while not done:
+                    data = h3_cache.read_object(h3_cache_bucket, h3_cache_object, offset)
+                    h3_cold.write_object(h3_cache_bucket, h3_cache_object, data, offset)
+                    
+                    done    = data.done
+                    offset += len(data)
+
+                # delete the CachedAt attribute
+                h3_cold.delete_object_metadata(h3_cache_bucket, h3_cache_object, "CachedAt")
+                
+                # remove the object from the cache
+                h3_cache.delete_object(h3_cache_bucket, h3_cache_object)
 
 def main(cmd=None):
     parser = argparse.ArgumentParser(description='Move Objects to Cold storage')
-    parser.add_argument('--cache_storage', required=True, help=f'Cache H3 storage URI')
-    parser.add_argument('--cold_storage', required=True, help=f'Cold H3 storage URI')
+    parser.add_argument('--cache_storage', required=True, help=f'Cache H3 storage URI.')
+    parser.add_argument('--cold_storage', required=True, help=f'Cold H3 storage URI.')
+    parser.add_argument('--expires_time', required=False, help=f'The maximum time that an object allowed to stay in the cache. (in seconds)')
     
     args = parser.parse_args(cmd)
     cache_config_path = args.cache_storage 
     cold_config_path = args.cold_storage
-
-    print(cache_config_path, cold_config_path)
+    expires_time = args.expires_time
 
     if cache_config_path and cold_config_path:
         h3_cache = pyh3lib.H3(cache_config_path)
         h3_cold  = pyh3lib.H3(cold_config_path)
 
-        # Move the objects from the hot storage to the cold storage   
-        MoveToCold(h3_cache, h3_cold)
+        if not expires_time:
+            expires_time = DEFAULT_EXPIRES_AT_TIME
+
+        # Move the objects from the cache storage to the cold storage   
+        MoveToCold(h3_cache, h3_cold, expires_time)
     else:
         parser.print_help(sys.stderr)
         sys.exit(1)
